Allows multiple dictionary keys in non-strict mode
Some PDFs break the specification by containing multiple definitions of
the same key in a single dictionary. This assumes the last definition is
the valid one, and allows to silently ignore previous definitions if
reading the PDF in non-strict mode.

Also for stream objects, in addition to searching one byte back for the
"endstream" marker this also adds support for searching one extra byte
forward. If found, the extra byte will be added to the stream data.